+++ mirgecom/phenolics/phenolics.py
@@ -71,3 +71,119 @@ def make_conserved(solid_species_mass, gas_density, gas_species_mass, energy):
     return WallConservedVars(solid_species_mass=solid_species_mass, energy=energy,
          gas_density=gas_density, gas_species_mass=gas_species_mass                    
     )
+
+
+@dataclass_array_container
+@dataclass(frozen=True)
+class PhenolicsDependentVars:
+    """State-dependent quantities for :class:`GasEOS`.
+
+    Prefer individual methods for model use, use this
+    structure for visualization or probing.
+
+    .. attribute:: temperature
+    .. attribute:: pressure
+    .. attribute:: velocity
+    """
+
+    temperature: DOFArray
+    pressure: DOFArray
+    velocity: DOFArray
+    progress: DOFArray
+
+    #XXX why are these nd.array and not DOFArray?
+    viscosity: np.ndarray
+    thermal_conductivity: np.ndarray
+    species_diffusivity: np.ndarray
+
+
+class PhenolicsEOS(metaclass=ABCMeta):
+    r"""Abstract interface to equation of state class.
+
+    .. automethod:: 
+    """
+
+    @abstractmethod
+    def pressure(self, cv: ConservedVars, temperature: DOFArray):
+        """Get the gas pressure."""
+
+    @abstractmethod
+    def temperature(self, cv: ConservedVars,
+                    temperature_seed: Optional[DOFArray] = None) -> DOFArray:
+        """Get the gas temperature."""
+
+    @abstractmethod
+    def gas_const(self, cv: ConservedVars):
+        r"""Get the specific gas constant ($R_s$)."""
+
+    @abstractmethod
+    def heat_capacity_cp(self, cv: ConservedVars, temperature: DOFArray):
+        r"""Get the specific heat capacity at constant pressure ($C_p$)."""
+
+    @abstractmethod
+    def viscosity(self, cv: ConservedVars) -> DOFArray:
+        r"""Get the gas dynamic viscosity, $\mu$."""
+
+    @abstractmethod
+    def thermal_conductivity(self, cv: ConservedVars) -> DOFArray:
+        r"""Get the gas thermal_conductivity, $\kappa$."""
+
+    @abstractmethod
+    def species_diffusivity(self, cv: ConservedVars) -> DOFArray:
+        r"""Get the vector of species diffusivities, ${d}_{\alpha}$."""
+
+    @abstractmethod
+    def emissivity(self, cv: ConservedVars, temperature: DOFArray):
+        r"""."""
+
+    @abstractmethod
+    def permeability(self, cv: ConservedVars, temperature: DOFArray):
+        r"""."""
+
+    @abstractmethod
+    def progress(self, cv: ConservedVars, temperature: DOFArray):
+        r"""."""
+
+    #FIXME
+    def dependent_vars(self, cv: ConservedVars,
+            temperature_seed: Optional[DOFArray] = None) -> PhenolicsDependentVars:
+        """Get an agglomerated array of the dependent variables."""
+        temperature = self.temperature(cv, temperature_seed)
+        return PhenolicsDependentVars(
+            progress=
+            temperature=temperature,
+            pressure=self.pressure(cv, temperature),
+            viscosity=
+            thermal_conductivity=
+            species_diffusivity=
+            emissivity=
+            permeability=
+        )
+
+
+class SimpleEquationOfState(PhenolicsEOS):
+
+    def __init__(self, composite):
+        self._degradation_model = composite
+
+    def heat_capacity_cp(self, cv, temperature):
+        return 
+
+    def gas_const(self, cv, temperature):
+        """Get specific gas constant R."""
+        return 
+
+    def pressure(self, cv, temperature):
+        return 
+
+    def permeability(self, cv, temperature):
+        return 
+
+    def emissivity(self, cv, temperature):
+        return 
+
+    def progress(self, cv, temperature):
+        return 
+
+
+
